[node] fix: "logging" location + refactoring
