+++ packages/node/src/local/websocket.ts
@@ -4,7 +4,7 @@ import type { WebSocketMessage } from 'js-proxy-shared/WebSocketMessage'
 import { parseWebSocketMessage, WebSocketMessageType } from 'js-proxy-shared/WebSocketMessage'
 import type { InstanceOptions } from '../utils/Options'
 import { WEBSOCKET_ROOT } from 'js-proxy-shared/constants'
-import { createErrorHandler, createErrorHandlerFor } from '../../../client/src/utils/logging'
+import { createErrorHandler, createErrorHandlerFor } from '../utils/logger'
 import { createErrorMessage } from '../utils/ws-messages'
 import type { ProxyState } from 'js-proxy-shared/ProxyState'
 import { listCertificates } from '../utils/cert-utils'
