+++ packages/node/src/proxy/request/proxy-request.ts
@@ -4,7 +4,7 @@ import * as https from 'https'
 import { getDecodedIncomingMessageData } from '../../utils/incoming-message'
 import type { ProxyRequest } from '../../server/ProxyRequest'
 import type { ProxyResponse } from '../../server/ProxyResponse'
-import { createErrorHandler } from '../../../../client/src/utils/logging'
+import { createErrorHandler } from '../../utils/logger'
 import { createRequestOptions, isReqHttps } from '../utils'
 import { cacheDir } from '../../utils/temp-dir'
 import { mkdirSync, writeFileSync } from 'fs'
