+++ packages/node/src/proxy/socket/proxy-request.ts
@@ -1,7 +1,7 @@
 import type { ProxyRequest } from '../../server/ProxyRequest'
 import type { Socket } from 'node:net'
 import type { ServerOptions } from '../../server'
-import { createErrorHandler } from '../../../../client/src/utils/logging'
+import { createErrorHandler } from '../../utils/logger'
 import * as utils from '../utils'
 import { isReqHttps, isReqWebSocket } from '../utils'
 import * as http from 'http'
@@ -9,7 +9,7 @@ import * as https from 'https'
 import { createErrorMessage } from '../../utils/ws-messages'
 import { WebSocketMessageType } from 'js-proxy-shared/WebSocketMessage'
 import { isLocalhost } from '../../utils/is-localhost'
-import { sendWsData } from '../../local/websocket'
+import { sendWsData } from '../../local'
 import { newLine } from './newline'
 import { NEWLINE } from '../constants'
 import { setupSocket } from './setup-socket'
