+++ packages/node/src/server/server.ts
@@ -1,7 +1,7 @@
 import type { IncomingMessage, Server, ServerResponse } from 'http'
 import * as http from 'http'
 import type { Logger } from '../utils/logger'
-import { clearScreen, createLogger } from '../utils/logger'
+import { clearScreen, createErrorHandler, createErrorHandlerFor, createLogger } from '../utils/logger'
 import * as https from 'https'
 import * as net from 'net'
 import type { AddressInfo } from 'ws'
@@ -15,7 +15,6 @@ import { isLocalhost } from '../utils/is-localhost'
 import { ProxyRequest } from './ProxyRequest'
 import { ProxyResponse } from './ProxyResponse'
 import { createErrorMessage, createProxyRequest } from '../utils/ws-messages'
-import { createErrorHandler, createErrorHandlerFor } from '../../../client/src/utils/logging'
 import { tempDir } from '../utils/temp-dir'
 import { join } from 'path'
 import { captureResponse } from '../utils/captureResponse'
